utils: remove sysmgr-needs script and add C programs
The previous shell script implementation had too much CPU overhead
because of the 'while true'. It also didn't exit properly. I have
decided to replace it with this small C program.

'util.c' is also added to make way for possible new programs to make
use of it. Perhaps a library could be created from this.